+++ test/suite/regressions/5943-session-start-notify.test.ts
@@ -105,7 +105,6 @@ type ReloadCommandContext = {
 		getShowHardwareCursor: () => boolean;
 		getClearOnShrink: () => boolean;
 	};
-	coalescingTerminal: { passNextFullRedraw: () => void };
 	keybindings: { reload: () => void };
 	customHeader?: unknown;
 	builtInHeader?: unknown;
@@ -179,7 +178,6 @@ function createReloadCommandContext(overrides: ReloadCommandContextOverrides = {
 			getClearOnShrink: () => false,
 			...overrides.settingsManager,
 		},
-		coalescingTerminal: { passNextFullRedraw: () => {} },
 		keybindings: { reload: () => {}, ...overrides.keybindings },
 		editorContainer: { clear: () => {}, addChild: () => {}, ...overrides.editorContainer },
 		ui: {